+++ OSDCloud/private/steps/5-drivers/step-Save-WindowsDriver-MSUpdate.ps1
@@ -14,7 +14,7 @@ function step-Save-WindowsDriver-MSUpdate {
         Write-Host -ForegroundColor DarkGray "[$(Get-Date -format s)] PowerShell 5.1 is required to run this step. Skip."
         return
     }
-    if (($IsVM -eq $true) -and ($global:OSDCloudDevice.OSDManufacturer -match 'Microsoft')) {
+    if (($IsVM -eq $true) -and ($global:OSDCoreDevice.OSDManufacturer -match 'Microsoft')) {
         Write-Host -ForegroundColor DarkGray "[$(Get-Date -format s)] Microsoft Update Drivers is not enabled for Microsoft Hyper-V. Skip."
         return
     }